fix(build): Export qemu_unregister_reset symbol on Windows
We introduced calls to unregister QEMU reset handlers in the HAL device
destructors. However, the `qemu_unregister_reset` symbol was not
exported in the Windows module definition file (`.def`), which would
cause a linker error on that platform.

This change adds `qemu_unregister_reset` to
`qemu-system-x86_64.def`, exporting the symbol and allowing the
HAL devices to correctly unregister their reset listeners on Windows.

Bug: 456299732
Change-Id: I3fd2a4745f894ce70c1005bc62d6c0af4cb8f8db