build: Raise macOS deployment target to 10.15
Bump the OSX_DEPLOYMENT_TARGET from 10.14 to 10.15 (Catalina).

This change is made to enable **C++20 language support**, which is required for
ongoing development. The target of 10.15 is the earliest macOS version that
fully supports the necessary toolchain features.

Note that this target increase may prevent approximately 0.04% of existing
users from running newer versions of the emulator, as their OS version will no
longer be supported.

Change-Id: I3fef1675b66746a9e02bbe4990fe85646256053a