hw/i3c: target: Add CCC is supported function
This is used in direct CCCs, when the controller re-addresses the target
after doign the broadcast address and CCC.
Per the spec, this is the only point where the target can NACK, so we
should provide a helper function to make organizing this easier.

This also helps simplify things in remote I3C, since the remote I3C flow
will more closely line up with reality.

As part of this, there will need to be changes on the g3 side. There
shouldn't be any visible changes to FW.

Tested: http://sponge2/a461be3f-41ed-436c-beed-909acef985f9
Google-Bug-Id: 458463163, 380273887
Change-Id: I4866a3c82dcddc1b1438f427dbd2d1695df00603
Signed-off-by: Joe Komlodi <komlodi@google.com>