snapshot: save and load rutabaga device
1, added an extra parameter to tell where to save snapshot
this is not the best solution so far, as it may not get accepted
by upstream, but a solution to get us started. we can re-visit
later, such as serialize the directory into QEMUFile*, but it
could make snapshot save and load slower, depends on how fast
we can ser/deserialize. An example of this path name is:
/path-to-avd-folder/snapshots/renderer/
rutabaga will create
rutabaga-2026-Apr-03-09-59-04 underneath, and this dir name is saved
to the snapshot blob in the qcow2 file associated with <snapshot_name>
(default is default_boot, as used in qemu2)
when the avd folder is deleted, this will also be deleted.

TODO:
note, this needs further improvements

2, added vmsd for rutabaga device, mostly saving VirtIODevice
and resources, and then during load, re-map those resouces
from guest (not change) to host (they do change for each load)

Bug: 496647016

Change-Id: Iff3d9bbbb32c4d3eeb8c9d686f70d5a1c58f3eca