Warn users on incorrect glasses emulator boot order
Adds a warning if the phone emulator is detected as started before the glasses emulator. The recommended boot order is glasses first, then phone, aligning with the pairing wizard UI. This is a stopgap solution to mitigate connection issues (b/435254006) caused by changing MAC addresses, crucial for the upcoming release.

This warning uses the port number as a proxy for mac address. This heuristic covers most common cases but has known limitations:
-   Doesn't handle port numbers hardcoded via command line.
-   Doesn't address multiple phone or glasses AVD scenarios.

Despite these limitations, this provides valuable guidance where none existed. A more robust fix is deferred due to time constraints.

For a screenshot of the popup, see b/461816199

Test: Lauch glasses AVD first: no popup
Test: Launch phone then glases: see the popup
Fixes: b/461816199
Bug: b/435254006
Change-Id: Icbe350127c3af9ddf3ef651d01a4d53aa38c6fd0