Allow battery to be zero, except on first load
We were previously setting the battery level to at least 1 every time we
stored and loaded the value, but also loading it every time visibility
changed, leading to strange behavior.

This allows the value to persist as 0, but changes it to 1 upon the
initial load, leading to more predictable behavior.

Bug: 464178898
Test: Boot emulator after shutting it down with battery level 0,
      confirm emulator boots up with battery level at 1.
Change-Id: Ibcdc073b5165ba8e9b1907036e2c39dfa05cc53c