hw/i3c/core: Send STOPs to all targets
Previously we would only send STOPs to targets participating on the bus,
and only send STOPs to all targets if it was ENTDAA. However, with
direct CCCs being modeled more accurately, we need to send STOPs in that
case as well.

We can just unconditionally send STOPs in all cases to handle this, and
targets should be able to handle spurious STOPs regardless.

Google-Bug-Id: 458463163
Change-Id: I2e3c3214557e0ca910f6bc6023356fad1052749d
Signed-off-by: Joe Komlodi <komlodi@google.com>