Ensure gles and gpu preferences are save per-avd
UI options are using a mixed way of getting and setting user
preferences, and this leads to choosing a gpu mode from the UI but not
seeing it affecting the gpu selection after a restart. Always save gpu
options into per-avd configuration file as the issues leading the user
to select software mdoe might be related to AVD's API level or
requirements.

Bug: 462074009
Test: manual, choose 'software' gpu mode from UI, check gles backend selected
Change-Id: I7539abdf9bb82f2141ce00841bd65ffaf882621f